fix: Mali_NativeWindowType must use unsigned short (not int) The Mali EGL fbdev driver reads width/height as 2-byte unsigned short each. Using 'int' (4 bytes) made the Mali blob interpret the bytes wrong — width became a huge number, height became 0, causing eglCreateWindowSurface to fail.
This matches the system's fbdev_window struct in EGL/fbdev_window.h

+++ src/video/mali-fbdev/SDL_malivideo.h
@@ -39,8 +39,8 @@ misrepresented as being the original software.
 /* Tipo nativo de janela Mali FBDEV */
 typedef struct
 {
-    int width;
-    int height;
+    unsigned short width;
+    unsigned short height;
 } Mali_NativeWindowType;
 
 /* Dados internos associados a cada SDL_VideoDisplay */
